% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningFunctions.R
\name{add_bin_features}
\alias{add_bin_features}
\title{Add bin centers for selected variables}
\usage{
add_bin_features(dat, bin_features, bin_type = "both", nbins)
}
\arguments{
\item{dat}{Data that will have bin centers added}

\item{bin_features}{Variables to bin on, must be numeric}

\item{bin_type}{"standard", "quantile", "both"}

\item{nbins}{The number of bins to create}
}
\value{
Return the training data with bin centers added.
}
\description{
Add bin centers for selected variables
}
