% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{bin_dictator_weighted}
\alias{bin_dictator_weighted}
\title{Establish bin dictator weights}
\usage{
bin_dictator_weighted(bin_model_weights)
}
\arguments{
\item{bin_model_weights}{nXM matrix of weights from the bin_weighted function}
}
\value{
matrix of bin-dictator weights
}
\description{
Take bin weights from bin_weights function and set all but best model weight in each bin to zero
}
