% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningFunctions.R
\name{bin_nd}
\alias{bin_nd}
\title{Make multidimensional binned data or bin definitions (for independent dimensions)}
\usage{
bin_nd(data, bin_features, nbins, bin_type = "standard",
  output = "definition")
}
\arguments{
\item{data}{data used to define bins}

\item{bin_features}{The features to bin on.  Must be two!}

\item{nbins}{The number of bins to use}

\item{bin_type}{The type of bin to use}
}
\value{
nd standard binning definition
}
\description{
From provided bin features, create multidimensional bin definition
}
\examples{
bin_nd(data=iris, bin_features=c("Petal.Length","Petal.Width","Sepal.Width"), nbins=c(2,3,4), bin_type="quantile", output="bin_data") 
bin_nd_def <- bin_nd(iris, c("Petal.Length","Petal.Width","Sepal.Width"),nbins=c(2,3,4), bin_type="standard", output="definition")
bin_nd(iris, c("Petal.Length","Petal.Width","Sepal.Width"), c(2,3,4), bin_type="standard", output="both")
}
