% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningFunctions.R
\name{bin_nd_by_def}
\alias{bin_nd_by_def}
\title{Bin new data based on n-dimensional binning definition}
\usage{
bin_nd_by_def(new_data, bin_nd_def)
}
\arguments{
\item{new_data}{data being binned}

\item{bin_nd_def}{binning definition based on the bin_nd function}
}
\description{
From provided bin features, create multidimensional bin definition
}
