% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningFunctions.R
\name{bin_test_by_train}
\alias{bin_test_by_train}
\title{Bin test data}
\usage{
bin_test_by_train(train_data, test_data, bin_features, bin_type, nbins)
}
\arguments{
\item{train_data}{Training data}

\item{test_data}{Test data}
}
\description{
Bin test data based on parameters defined by training data
this will be used to assign values for test data sets to bins defined by training data
}
