% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{bin_weighted}
\alias{bin_weighted}
\title{Model weights for "weight_type == "bin weighted" VERSION 2}
\usage{
bin_weighted(bin_features, bin_type, nbins, train_data_preds, test_data, M, K)
}
\arguments{
\item{bin_features}{Training data with predicted class columns from each model \code{1,...,M}}

\item{bin_type}{Type of binning {"standard","quantile","iterative quantile"}}

\item{nbins}{vector containing number of bins in each dimension}

\item{train_data_preds}{data frame containing training data and CV prediction columns}

\item{test_data}{data frame containing test data. Must have same column names as training data}

\item{M}{number of models in bin weighted ensemble}

\item{K}{number of true classes}
}
\value{
matrix of bin weights
}
\description{
Calculate the model weights when "weight_type" == "bin weighted"
# TODO this needs to be cleaned up
}
