% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFunctions.R
\name{cv_cohorts}
\alias{cv_cohorts}
\title{CV cohort generator}
\usage{
cv_cohorts(data_size, cv_K)
}
\arguments{
\item{data_size}{Sample size in data being cross validated}

\item{cv_k}{Number of Folds in CV. Default=10}
}
\value{
\code{preds}
}
\description{
Function for making CV fold indeces
}
