% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFunctions.R
\name{cv_preds}
\alias{cv_preds}
\title{K-fold CV Predictions}
\usage{
cv_preds(model_type, dat, true_classes, cv_k = 10)
}
\arguments{
\item{true_classes}{Vector holding the order of the true labels}

\item{cv_k}{Number of Folds in CV. Default=10}

\item{mod}{An RWeka model object}

\item{data}{A data frame to predict}
}
\value{
\code{preds}
}
\description{
Function for making K-fold CV predictions of classes using member models from RWeka
}
