% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{cv_testing_all_ensembles}
\alias{cv_testing_all_ensembles}
\title{cv_testing_all_ensembles}
\usage{
cv_testing_all_ensembles(data, model_types, bin_features_all, nbins_all,
  equal_bins = FALSE, cv_K = 10)
}
\arguments{
\item{train_data}{Data to}

\item{test_data}{Data to predict with the ensemble}

\item{model_list}{List of RWeka models fit to train_data}
}
\description{
cv_testing_all_ensembles
}
