% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{eval_cv_test_results}
\alias{eval_cv_test_results}
\title{eval_cv_test_results}
\usage{
eval_cv_test_results(cv_test_results_list)
}
\arguments{
\item{train_data}{Data to}
}
\description{
eval_cv_test_results
}
