% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFunctions.R
\name{eval_ensemble_members}
\alias{eval_ensemble_members}
\title{Function to evaluated the ensemble test accuracy against member accuracies}
\usage{
eval_ensemble_members(model_list, test_data)
}
\arguments{
\item{model_list}{A list of RWeka Models}

\item{test_data}{Data to predict with the ensemble}
}
\value{
Predictions for the provided data
}
\description{
Function to evaluated the ensemble test accuracy against member accuracies
}
