% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{iq_binned_testing}
\alias{iq_binned_testing}
\title{binned_testing}
\usage{
iq_binned_testing(train_data, test_data, model_list, bin_features_list,
  nbins_list, true_classes)
}
\arguments{
\item{train_data}{Data to}

\item{test_data}{Data to predict with the ensemble}

\item{model_list}{List of RWeka models fit to train_data}
}
\description{
binned_testing
}
