% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{knn_weighted}
\alias{knn_weighted}
\title{Model weights for "weight_type == "knn"}
\usage{
knn_weighted(train_data, test_data, M, scale = TRUE, knn_size = 10)
}
\arguments{
\item{train_data}{Training data with predicted class columns from each model \code{1,...,M}}

\item{scale}{TRUE/FALSE for rescaling data (default: scale=TRUE)}

\item{n}{The number of instances in the test data}

\item{neighbors}{number of neighbors used in accuracy estimates for each model}
}
\value{
matris of weights nrow=number of test points, rcol=M=number of models
}
\description{
Calculate the model weights using k-nearest neighbor accuracy weights when "weight_type" == "knn"
}
