% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{make_bin_feature_list_pairs}
\alias{make_bin_feature_list_pairs}
\title{make_bin_feature_list_pairs}
\usage{
make_bin_feature_list_pairs(bin_features_all, ordered = FALSE)
}
\arguments{
\item{train_data}{Data to}

\item{test_data}{Data to predict with the ensemble}

\item{model_list}{List of RWeka models fit to train_data}
}
\description{
make_bin_feature_list_pairs
}
