% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFunctions.R
\name{make_ensemble}
\alias{make_ensemble}
\title{Build a weighted ensemble list}
\usage{
make_ensemble(train_preds = NULL, model_list = NULL, weightType = NULL,
  comb_rule = NULL, bin_type = "standard", bin_features = NULL,
  nbins = NULL, knn_size = 10)
}
\arguments{
\item{model_list}{A list of pre trained models (using the same data as \code{train})}

\item{weightType}{How the ensemble should be weighted; "unweighted", "weighted", "bin weighted"}

\item{comb_rule}{How the ensemble predictions should be comined; "average posterior","majority vote"}

\item{bin_type}{How bins should be created when \code{weightType} is "bin weighted"; "average posterior","majority vote"}

\item{bin_features}{The name of two numeric bins when \code{weightType} is "bin weighted"}

\item{nbins}{The number of bins to create when \code{weightType} is "bin weighted"}

\item{train_data}{Data on which to train the ensemble}
}
\value{
A trained ensemble that can be used to predict new data points
}
\description{
Gather all parameters of a weighted ensemble of RWeka models into formated list
}
