% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningFunctions.R
\name{make_feature_pair_df}
\alias{make_feature_pair_df}
\title{Create feature pair data frame}
\usage{
make_feature_pair_df(bin_features, bin_type, nbins)
}
\arguments{
\item{bin_features}{The features to bin on.  Must be two!}

\item{bin_type}{The type of bin to use}

\item{nbins}{The number of bins to use}
}
\value{
\code{featurePairs}
}
\description{
From provided feature pairs, create a data frame that holds additional information.
}
