% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFunctions.R
\name{make_model_list}
\alias{make_model_list}
\title{Function for building model list given Weka names}
\usage{
make_model_list(model_types, data, ...)
}
\arguments{
\item{model_types}{A vector of Weka model type names}

\item{data}{A data frame to predict}
}
\value{
list of weka models
}
\description{
Function for making predictions of classes using member models.
}
