% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{make_model_metric_array}
\alias{make_model_metric_array}
\title{Model metric array for combination rules}
\usage{
make_model_metric_array(combination_rule, model_storage_list, test_data,
  true_classes)
}
\arguments{
\item{combination_rule}{"majority vote" or "average posterior"}

\item{model_storage_list}{A list holding models from RWeka}

\item{test_data}{A data frame holding data on which to test}

\item{true_classes}{An array holding the order of the true labels CANDIDATE FOR REPLACEMENT}
}
\value{
\code{model_metric}
}
\description{
Function for making model metric array for combination rules\cr
I DO NOT YET UNDERSTAND THIS ONE \cr
NEEDS TO BE GENERALIZED
}
\examples{

}
