% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{make_model_metric_array_avgpost}
\alias{make_model_metric_array_avgpost}
\title{Average posterior vote}
\usage{
make_model_metric_array_avgpost(test_post_probs)
}
\arguments{
\item{test_post_probs}{Posterior probabilities from model predictions on test data}
}
\value{
\code{model_metric}
}
\description{
Create ensemble combination metric arrays with dimension \code{n X K X M}
with elements v_lim = obs i metric for class l from member m
Using average posterior rule\cr\cr

\code{n} is the number of observations\cr
\code{K} is the number of classes\cr
\code{M} is the number of models\cr

I DO NOT YET UNDERSTAND THIS ONE
}
\examples{

}
