% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{make_model_metric_array_majvote}
\alias{make_model_metric_array_majvote}
\title{Majority vote}
\usage{
make_model_metric_array_majvote(test_preds)
}
\arguments{
\item{test_post_probs}{Posterior probabilities from model predictions on test data}
}
\value{
\code{model_metric}
}
\description{
Create ensemble combination metric arrays with dimension n X K X M
with elements v_lim = obs i metric for class l from member m
Using majority rule

I DO NOT YET UNDERSTAND THIS ONE
}
\examples{

}
