% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{make_nbins_list_pairs}
\alias{make_nbins_list_pairs}
\title{make_nbins_list_pairs}
\usage{
make_nbins_list_pairs(nbins_all, equal_bins = FALSE)
}
\arguments{
\item{train_data}{Data to}

\item{test_data}{Data to predict with the ensemble}

\item{model_list}{List of RWeka models fit to train_data}
}
\description{
make_nbins_list_pairs
}
