% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFunctions.R
\name{make_preds}
\alias{make_preds}
\title{Predict classes using member models}
\usage{
make_preds(data, model_list, true_classes)
}
\arguments{
\item{data}{A data frame to predict}

\item{model_list}{A list holding RWeka models}

\item{true_classes}{Array holding the order of the true labels}
}
\value{
\code{preds}
}
\description{
Function for making predictions of classes using member models.
}
