% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binningFunctions.R
\name{make_train_bins_index}
\alias{make_train_bins_index}
\title{Make bin index}
\usage{
make_train_bins_index(train_data, featurePairs, bin_type = "standard", nbins)
}
\arguments{
\item{train_data}{Training data from which to build bins}

\item{featurePairs}{Data frame with one row that has original binning features' names, and then augmented feature names}

\item{bin_type}{"standard", "quantile"}

\item{nbins}{The number of bins}
}
\description{
make bin index set for ALL POSSIBLE bins in training data
featPair <- featurePairs[1,] TODO NEED TO FIX INDEXING FOR BINS, CANNOT SKIP STORAGE OF BIN INDECES THAT ARE EMPTY SINCE THEY ARE NEEDED FOR NEW OBSERVATIONS
featPair <- featurePairs[1,]   FEATURE PAIRS ONLY HOLDS ONE NOW
}
