% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFunctions.R
\name{make_train_preds}
\alias{make_train_preds}
\title{Make training predictions from all models in ensemble}
\usage{
make_train_preds(train_data, model_list, true_classes)
}
\arguments{
\item{train_data}{Data on which to train the ensemble}

\item{model_list}{A list of pre trained models (using the same data as \code{train})}

\item{true_classes}{vector of true classes}
}
\value{
A trained ensemble that can be used to predict new data points
}
\description{
Make training predictions from all models in ensemble
}
