% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFunctions.R
\name{predictEnsemble}
\alias{predictEnsemble}
\title{Predict using an ensemble classifier}
\usage{
predictEnsemble(ensemble, test_data)
}
\arguments{
\item{ensemble}{An ensemble classifier built with \code{buildWeightedEnsemble}}

\item{test_data}{Data to predict with the ensemble}
}
\value{
Predictions for the provided data
}
\description{
Predict using an ensemble classifier created from this package
}
