% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFunctions.R
\name{RA}
\alias{RA}
\title{Function to calculate classifier accuracy based on confusion matrix}
\usage{
RA(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{confusion matrix from a classifier}
}
\value{
correct classification rate (numeric vector of length 1)
}
\description{
Function to calculate classifier accuracy based on confusion matrix
}
