% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{testing_all_ensembles}
\alias{testing_all_ensembles}
\title{testing_all_ensembles}
\usage{
testing_all_ensembles(train_data, test_data, model_types, bin_features_all,
  nbins_all, equal_bins = FALSE)
}
\arguments{
\item{train_data}{Data to}

\item{test_data}{Data to predict with the ensemble}

\item{model_list}{List of RWeka models fit to train_data}
}
\description{
testing_all_ensembles
}
