% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudyFunctions.R
\name{unbinned_testing}
\alias{unbinned_testing}
\title{Exhaustively testing ensemble on new dataset}
\usage{
unbinned_testing(train_data, test_data, model_list, true_classes)
}
\arguments{
\item{train_data}{Data to}

\item{test_data}{Data to predict with the ensemble}

\item{model_list}{List of RWeka models fit to train_data}

\item{weight_type}{vector containing bin weight types from {"bin weighted","bin dictator"}}

\item{comb_rule}{vector containing combination rules from {"majority vote","average posterior"}}

\item{bin_type}{vector containing bin types from {"standard","quantile","iterative quantile"}}

\item{bin_features}{vector containing names of columns to bin}

\item{nbins}{vector containing number of bins in each dimension for bin_features (will use all combinations of these values)}
}
\value{
test accuracy rates for all weighting/voting/binning specifications requested
}
\description{
Predict using an ensemble classifier created from this package
}
