% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightingFunctions.R
\name{weighted}
\alias{weighted}
\title{Model weights for "weight_type == "weighted"}
\usage{
weighted(train_data, M, n)
}
\arguments{
\item{train_data}{Training data with predicted class columns from each model \code{1,...,M}}

\item{n}{The number of instances in the test data}
}
\value{
matris of weights
}
\description{
Calculate the model weights when "weight_type" == "weighted"
}
